Added error return to timestamp functions
This change is made to make the functions easier to test and reason with.
It also required changes in warcserver sorter. Behaviour should remain the same as
we ignore the error case there while the non error value should be the same as before.

Of cource this is not optimal and should be refactored in a future commit to be more
idiomatic.