feat: write WARC-Date with up to nanosecond precision
This commit increases the precision of the WARC-Date field for WARC 1.1
records up to 9 decimal fractions of a second (nanosecond) . In practice the
clock resolution of unix systems is ~1ms and ~0.5ms for windows systems.

WARC 1.0 records will be written with second precision as was the case for
all records written with this library until now.

The WARC 1.1 specification says the WARC-Date field may be specified in any
levels of granularity according to the [W3CDTF]. It also says WARC-Date
should be specified with as much precision as is accurately known.

See https://iipc.github.io/warc-specifications/specifications/warc-format/warc-1.1/#warc-date-mandatory
See https://www.w3.org/TR/NOTE-datetime
See https://pkg.go.dev/time#hdr-Timer_Resolution